requirements.txt
pylint
mypy
# numpy==1.23.5
# opencv-python==4.5.5.64
Pillow==9.3.0
opennsfw2~=0.10.2
fastapi~=0.108.0
uvicorn~=0.25.0
python-multipart
numpy~=1.23.5
tensorflow